+++ MCXLatticeMatter/test/runtests.jl
@@ -1,6 +1,6 @@
 using Test
 
-@testset "LatticePolymerSystems" begin
+@testset "MCXLatticeMatter" begin
     include("test_lattice_polymer.jl")
     include("test_mc_updates.jl")
 end