+++ MCXSpins/test/runtests.jl
@@ -1,9 +1,9 @@
 using Test
 using Random
 using MonteCarloX
-using SpinSystems
+using MCXSpins
 
-@testset "SpinSystems" begin
+@testset "MCXSpins" begin
     include("test_ising.jl")
     include("test_blume_capel.jl")
 end